+++ sqlearn/__init__.py
@@ -1,4 +1,4 @@
 # Importing classes from modules (filenames) to make user imports easier
 
-from .LinearRegressor import LinearRegressor
-from .LogisticRegressor import LogisticRegressor
+from .linear_regression import LinearRegressor
+from .logistic_regression import LogisticRegressor
