+++ sqlearn/logistic_regression.py
@@ -1,17 +1,23 @@
 import numpy as np
 import pandas as pd
-from .Model import Model
-import random
+from .sgd_regression import SGDRegressor
 
-class SGDRegressor(Model):
+def sigmoid(x):
+    ex = np.exp(x)
+    return ex / (1 + ex)
+
+class LogisticRegressor(SGDRegressor):
     """
-    A parent class for a regression model that's trained using stochastic 
-    gradient descent. 
-    
-    The functions calc_gradient and predict are not implemented.
+    A class that represents a trainable logistic regression binary 
+    classifier.
+    Uses average negative log likelihood as the loss function.
 
     Attributes:
     ----------
+    thresh (default=0.5)
+        The threshold at which predictions from the logistic regressor
+        change from 0 to 1. If thresh=None, then the model will output
+        probabilities instead of classifications.
     epochs (default=5)
         The number of passes the model makes through the dataset.
     batch_size (default=1)
@@ -25,25 +31,29 @@ class SGDRegressor(Model):
     
     Functions:
     ----------
-    calc_gradient(self, batch_X, batch_y) [NOT IMPLEMENTED]
+    calc_gradient(self, batch_X, batch_y)
         A function that calculates the gradient of the loss function.
     sgd(self, batch_X, batch_y):
         Performs one step of stochastic gradient descent.
     train(self, X, y):
         Trains the model on a dataset and the corresponding labels.
-    predict(self, X): [NOT IMPLEMENTED]
+    predict(self, X):
         Makes prediction from a dataframe of observations.
-
     """
 
-    def __init__(self, epochs=5, batch_size=1, learning_rate=0.001, seed=0):
+    def __init__(self, thresh=0.5, epochs=5, batch_size=1, learning_rate=0.001, seed=0):
         """
-        Initializes the SGDRegressor.
+        Initializes the LogisticRegressor.
 
         Parameters:
         ----------
+        thresh (default=0.5)
+            The threshold at which predictions from the logistic regressor 
+            change from 0 to 1. If thresh=None, then the model will output
+            probabilities instead of classifications.
         epochs (default=5)
             The number of passes the model makes through the dataset.
+            
         batch_size (default=1)
             The number of observations used to calculate the gradient during
             the process of stochastic gradient descent.
@@ -53,11 +63,8 @@ class SGDRegressor(Model):
         seed (default=0)
             The seed used for random processes.
         """
-        self.epochs = epochs
-        self.batch_size = batch_size
-        self.learning_rate = learning_rate
-        self.seed = seed
-        random.seed(self.seed)
+        self.thresh = thresh
+        super().__init__(epochs, batch_size=batch_size, learning_rate=learning_rate, seed=seed)
 
     def calc_gradient(self, batch_X, batch_y):
         """
@@ -69,32 +76,21 @@ class SGDRegressor(Model):
             A dataframe of a batch of observations without labels
         batch_y
             A dataframe (or series) of the labels of those observations.
-
+        
         Returns:
         ----------
         gradient
             The calculated gradient.
         """
-        pass
-
-    def sgd(self, batch_X, batch_y):
-        """
-        Performs one step of stochastic gradient descent.
-
-        Parameters:
-        ----------
-        batch_X
-            A dataframe containing a batch of observations.
-        batch_y
-            A dataframe (or series) containing the labels
-            for batch_X
-        """
-        gradient = self.calc_gradient(batch_X, batch_y)
-        self.theta = self.theta - self.learning_rate*gradient
+        gradient = 0
+        for i, x in batch_X.iterrows():
+            gradient += (sigmoid(x @ self.theta) - batch_y[i]) * x
+        gradient /= batch_X.shape[0]
+        return gradient
 
     def train(self, X, y):
         """
-        Trains the SGDRegressor using stochastic gradient descent and the
+        Trains the LogisticRegressor using stochastic gradient descent and the
         initialized hyperparameters.
 
         Parameters:
@@ -106,18 +102,7 @@ class SGDRegressor(Model):
             A dataframe (or series) containing the labels for each
             observation in X.
         """
-        ones = pd.DataFrame(np.ones((X.shape[0], 1)))
-        X_ones = pd.concat((ones, X), axis=1)
-        m = X_ones.shape[0]
-        n = X_ones.shape[1]
-        self.theta = pd.Series(np.zeros(n))
-        self.theta = self.theta.set_axis(X_ones.columns)
-        for epoch in range(self.epochs):
-            for batch in range(m):
-                batch_i = random.sample(range(m), self.batch_size)
-                batch_X = X_ones.iloc[batch_i]
-                batch_y = y.iloc[batch_i]
-                self.sgd(batch_X, batch_y)
+        super().train(X, y)
 
     def predict(self, X):
         """
@@ -133,4 +118,9 @@ class SGDRegressor(Model):
         y
             The predicted labels of the observations.
         """
-        pass
+        ones = pd.DataFrame(np.ones((X.shape[0], 1)))
+        X_ones = pd.concat((ones, X), axis=1)
+        preds = sigmoid(X_ones @ self.theta)
+        if self.thresh != None:
+            preds = preds >= self.thresh
+        return preds
