+++ sqlearn/sgd_regression.py
@@ -1,23 +1,17 @@
 import numpy as np
 import pandas as pd
-from .SGDRegressor import SGDRegressor
+from .model_interface import Model
+import random
 
-def sigmoid(x):
-    ex = np.exp(x)
-    return ex / (1 + ex)
-
-class LogisticRegressor(SGDRegressor):
+class SGDRegressor(Model):
     """
-    A class that represents a trainable logistic regression binary 
-    classifier.
-    Uses average negative log likelihood as the loss function.
+    A parent class for a regression model that's trained using stochastic 
+    gradient descent. 
+    
+    The functions calc_gradient and predict are not implemented.
 
     Attributes:
     ----------
-    thresh (default=0.5)
-        The threshold at which predictions from the logistic regressor
-        change from 0 to 1. If thresh=None, then the model will output
-        probabilities instead of classifications.
     epochs (default=5)
         The number of passes the model makes through the dataset.
     batch_size (default=1)
@@ -31,29 +25,25 @@ class LogisticRegressor(SGDRegressor):
     
     Functions:
     ----------
-    calc_gradient(self, batch_X, batch_y)
+    calc_gradient(self, batch_X, batch_y) [NOT IMPLEMENTED]
         A function that calculates the gradient of the loss function.
     sgd(self, batch_X, batch_y):
         Performs one step of stochastic gradient descent.
     train(self, X, y):
         Trains the model on a dataset and the corresponding labels.
-    predict(self, X):
+    predict(self, X): [NOT IMPLEMENTED]
         Makes prediction from a dataframe of observations.
+
     """
 
-    def __init__(self, thresh=0.5, epochs=5, batch_size=1, learning_rate=0.001, seed=0):
+    def __init__(self, epochs=5, batch_size=1, learning_rate=0.001, seed=0):
         """
-        Initializes the LogisticRegressor.
+        Initializes the SGDRegressor.
 
         Parameters:
         ----------
-        thresh (default=0.5)
-            The threshold at which predictions from the logistic regressor 
-            change from 0 to 1. If thresh=None, then the model will output
-            probabilities instead of classifications.
         epochs (default=5)
             The number of passes the model makes through the dataset.
-            
         batch_size (default=1)
             The number of observations used to calculate the gradient during
             the process of stochastic gradient descent.
@@ -63,8 +53,11 @@ class LogisticRegressor(SGDRegressor):
         seed (default=0)
             The seed used for random processes.
         """
-        self.thresh = thresh
-        super().__init__(epochs, batch_size=batch_size, learning_rate=learning_rate, seed=seed)
+        self.epochs = epochs
+        self.batch_size = batch_size
+        self.learning_rate = learning_rate
+        self.seed = seed
+        random.seed(self.seed)
 
     def calc_gradient(self, batch_X, batch_y):
         """
@@ -76,21 +69,32 @@ class LogisticRegressor(SGDRegressor):
             A dataframe of a batch of observations without labels
         batch_y
             A dataframe (or series) of the labels of those observations.
-        
+
         Returns:
         ----------
         gradient
             The calculated gradient.
         """
-        gradient = 0
-        for i, x in batch_X.iterrows():
-            gradient += (sigmoid(x @ self.theta) - batch_y[i]) * x
-        gradient /= batch_X.shape[0]
-        return gradient
+        pass
+
+    def sgd(self, batch_X, batch_y):
+        """
+        Performs one step of stochastic gradient descent.
+
+        Parameters:
+        ----------
+        batch_X
+            A dataframe containing a batch of observations.
+        batch_y
+            A dataframe (or series) containing the labels
+            for batch_X
+        """
+        gradient = self.calc_gradient(batch_X, batch_y)
+        self.theta = self.theta - self.learning_rate*gradient
 
     def train(self, X, y):
         """
-        Trains the LogisticRegressor using stochastic gradient descent and the
+        Trains the SGDRegressor using stochastic gradient descent and the
         initialized hyperparameters.
 
         Parameters:
@@ -102,7 +106,18 @@ class LogisticRegressor(SGDRegressor):
             A dataframe (or series) containing the labels for each
             observation in X.
         """
-        super().train(X, y)
+        ones = pd.DataFrame(np.ones((X.shape[0], 1)))
+        X_ones = pd.concat((ones, X), axis=1)
+        m = X_ones.shape[0]
+        n = X_ones.shape[1]
+        self.theta = pd.Series(np.zeros(n))
+        self.theta = self.theta.set_axis(X_ones.columns)
+        for epoch in range(self.epochs):
+            for batch in range(m):
+                batch_i = random.sample(range(m), self.batch_size)
+                batch_X = X_ones.iloc[batch_i]
+                batch_y = y.iloc[batch_i]
+                self.sgd(batch_X, batch_y)
 
     def predict(self, X):
         """
@@ -118,9 +133,4 @@ class LogisticRegressor(SGDRegressor):
         y
             The predicted labels of the observations.
         """
-        ones = pd.DataFrame(np.ones((X.shape[0], 1)))
-        X_ones = pd.concat((ones, X), axis=1)
-        preds = sigmoid(X_ones @ self.theta)
-        if self.thresh != None:
-            preds = preds >= self.thresh
-        return preds
+        pass
